\name{igraph-class}
\Rdversion{1.1}
\docType{class}
\alias{igraph-class}

\title{Class \code{"igraph"}}
\description{
	igraph object defines network.
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Slots}{
  \describe{
    \item{\code{.S3Class}:}{Object of class \code{"character"} }
  }
}
\section{Extends}{
Class \code{"\linkS4class{oldClass}"}, directly.
}
\section{Methods}{
No methods defined with class "igraph" in the signature.
}
\author{
	Shailesh Tripathi
}


\seealso{
	igraph
}
\examples{
showClass("igraph")
}
\keyword{classes}
